refactor(harness): freeze gate entrypoint facade
Work-Run: b178eef8-09c6-4c1a-9cb8-ebf0be65a2ba
Work-Receipt: g0-r0
