+++ scripts/harness/__tests__/classify-changed-paths.test.mjs
@@ -513,7 +513,7 @@ describe('CI capability wiring', () => {
     expect(workflow).toContain('pnpm lint:affected');
     expect(workflow).toContain('pnpm examples:typecheck:affected');
     expect(workflow).toContain('pnpm build\n');
-    expect(workflow).toContain('start_check test pnpm test\n');
+    expect(workflow).toContain('start_check test env HOME="$empty_home" pnpm test\n');
     expect(workflow).toContain('start_check typecheck pnpm typecheck\n');
     expect(workflow).toContain('start_check lint pnpm lint\n');
     expect(workflow).toContain('pnpm examples:typecheck\n');
