+++ scripts/harness/__tests__/merge-gate-disposition.test.mjs
@@ -175,7 +175,8 @@ function stubbedPath(prs) {
       '    mine = mine.filter((c) => /actionable findings:\\s*[0-9]+/i.test(c.body));',
       '  }',
       '  mine.sort((a, b) => (a.at < b.at ? -1 : a.at > b.at ? 1 : 0));',
-      '  console.log(JSON.stringify(mine.at(-1) ?? {}));',
+      '  // The hook counts the complete verdict list before selecting its newest entry.',
+      '  console.log(JSON.stringify(mine));',
       '  process.exit(0);',
       '}',
       'if (args.includes("pr checks")) { process.exit(0); }',
