+++ scripts/harness/file-size-baseline.json
@@ -83,7 +83,7 @@
   "packages/dag-nodes/gemini-image-edit/src/runtime-core.ts": 318,
   "packages/dag-nodes/instant-node/src/index.ts": 574,
   "packages/dag-nodes/utility-text/src/index.ts": 626,
-  "scripts/harness/allocate-work-item-id.mjs": 472,
+  "scripts/harness/allocate-work-item-id.mjs": 479,
   "scripts/harness/check-agent-server-boundary.mjs": 728,
   "scripts/harness/check-background-workspace-conformance.mjs": 356,
   "scripts/harness/check-backlog-placement.mjs": 331,
@@ -101,18 +101,18 @@
   "scripts/harness/checkpoint-evidence-contract.mjs": 463,
   "scripts/harness/cleanup-drift.mjs": 541,
   "scripts/harness/commit-message-claims.mjs": 384,
-  "scripts/harness/gate.mjs": 2560,
+  "scripts/harness/gate-operations.mjs": 2279,
   "scripts/harness/github-api.mjs": 356,
   "scripts/harness/github-issue-triage.mjs": 1303,
   "scripts/harness/lessons-lib.mjs": 367,
   "scripts/harness/live-provider-smoke.mjs": 458,
   "scripts/harness/loop-run.mjs": 1010,
-  "scripts/harness/new-spec.mjs": 446,
+  "scripts/harness/new-spec.mjs": 449,
   "scripts/harness/promote.mjs": 346,
   "scripts/harness/record-local-review.mjs": 540,
   "scripts/harness/release-run.mjs": 544,
   "scripts/harness/review-change.mjs": 338,
-  "scripts/harness/run-all-scans.mjs": 1914,
+  "scripts/harness/run-all-scans.mjs": 1910,
   "scripts/harness/scan-action-references.mjs": 579,
   "scripts/harness/scan-aggregate-naming.mjs": 689,
   "scripts/harness/scan-api-pagination.mjs": 358,
@@ -122,7 +122,7 @@
   "scripts/harness/scan-composition-neutrality.mjs": 529,
   "scripts/harness/scan-consistency.mjs": 438,
   "scripts/harness/scan-dist-freshness.mjs": 387,
-  "scripts/harness/scan-guard-scope-fail-closed.mjs": 1436,
+  "scripts/harness/scan-guard-scope-fail-closed.mjs": 1442,
   "scripts/harness/scan-hook-catalog.mjs": 307,
   "scripts/harness/scan-hook-enforcement-reachable.mjs": 609,
   "scripts/harness/scan-interface-family-owner.mjs": 509,
@@ -150,5 +150,6 @@
   "scripts/harness/scan-user-execution-plan-order.mjs": 2415,
   "scripts/harness/scan-workflow-permissions.mjs": 388,
   "scripts/harness/shared.mjs": 839,
+  "scripts/harness/work-run-store.mjs": 337,
   "scripts/harness/worktree-gate.mjs": 431
 }
