+++ scripts/harness/gate-entrypoint-baseline.json
@@ -1,0 +1,5 @@
+{
+  "schemaVersion": 1,
+  "facadePath": "scripts/harness/gate.mjs",
+  "sha256": "1919263e1ad906ea662d8b56e7a584f81134f386ecc50d05dcbbdbd747cea09e"
+}
