+++ scripts/harness/scan-gate-evaluator-isolation.mjs
@@ -21,7 +21,10 @@ import { resolveWorkspaceRoot } from './shared.mjs';
 
 const ROOT = resolveWorkspaceRoot(import.meta);
 const EXAMINED = ['::', 'examined::'].join('');
-const EVALUATOR_PREFIXES = ['scripts/harness/gate.mjs', '.claude/hooks/'];
+const EVALUATOR_PREFIXES = [
+  'scripts/harness/gate-operations.mjs',
+  '.claude/hooks/',
+];
 const SPEC_PREFIX = '.agents/spec-docs/';
 
 function git(args) {
